// src/generator/dmmf/types.ts
export type ReadonlyDeep<O> = {
  +readonly [K in keyof O]: ReadonlyDeep<O[K]>;
};

export namespace DMMF {
// ...
  export type UniqueIndex = ReadonlyDeep<{
    name: string;
    fields: string[];
  }>;
  export type PrimaryKey = ReadonlyDeep<{
    name: string | null;
    fields: string[];
  }>;
  export type Model = ReadonlyDeep<{
    name: string;
    dbName: string | null;
    fields: ModelField[];
    uniqueFields: string[][];
    uniqueIndexes: UniqueIndex[];
    // documentation?: string;
    primaryKey: PrimaryKey | null;
    // isGenerated?: boolean;
    // additional props:
    typeName: string;
    docs: string | undefined;
    plural: string | undefined;
    isOmitted: { output: boolean };
  }>;
  export type FieldKind = "scalar" | "object" | "enum" | "unsupported";
  export type FieldNamespace = "model" | "prisma";
  export type FieldLocation =
    | "scalar"
    | "inputObjectTypes"
    | "outputObjectTypes"
    | "enumTypes"
    | "fieldRefTypes";
  // Field
  export type ModelField = ReadonlyDeep<{
    // kind: FieldKind;
    name: string;
    isRequired: boolean;
    isList: boolean;
    isUnique: boolean;
    isId: boolean;
    isReadOnly: boolean;
    isGenerated?: boolean;
    isUpdatedAt?: boolean;
    // type: string;
    dbNames?: string[] | null;
    hasDefaultValue: boolean;
    default?: FieldDefault | FieldDefaultScalar | FieldDefaultScalar[];
    relationFromFields?: string[];
    relationToFields?: string[];
    relationOnDelete?: string;
    relationName?: string;
    // documentation?: string;
    // [key: string]: any;
    // additional props:
    type: string;
    location: FieldLocation;
    typeFieldAlias?: string;
    typeGraphQLType: string;
    fieldTSType: string;
    docs: string | undefined;
    isOmitted: {
      input: boolean | InputOmitSetting[];
      output: boolean;
    };
    isOptional: {
      input: boolean | InputOmitSetting[];
      output: boolean;
    };
  }>;
// ...
  export type OutputType = ReadonlyDeep<{
    name: string;
    // fields: SchemaField[];
    // additional props:
    fields: OutputSchemaField[];
    typeName: string;
  }>;
// ...
  export type OutputSchemaField = SchemaField &
    ReadonlyDeep<{
      argsTypeName: string | undefined;
    }>;
// ...
}